@@ -1,12 +1,17 @@
+import 'dart:async';
+
 import 'package:confetti/confetti.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:fokus/logic/task_instance/task_instance_cubit.dart';
 import 'package:fokus/logic/timer/timer_cubit.dart';
-import 'package:fokus/model/currency_type.dart';
 import 'package:fokus/services/app_locales.dart';
-import 'package:fokus/services/ticker.dart';
+import 'package:fokus/utils/duration_utils.dart';
 import 'package:fokus/utils/theme_config.dart';
 import 'package:fokus/widgets/app_header.dart';
+import 'package:logging/logging.dart';
+import 'package:vibration/vibration.dart';
+
 
 import 'chips/attribute_chip.dart';
 import 'large_timer.dart';
@@ -18,9 +23,9 @@ class TaskAppHeader extends StatefulWidget with PreferredSizeWidget {
 	final Widget appHeaderWidget;
 	final String helpPage;
 	final Function breakPerformingTransition;
-	final bool isBreak;
+	final TaskInstanceProvider state;
 
-	TaskAppHeader({Key key, @required this.height, this.title, this.text, this.appHeaderWidget, this.helpPage,this.breakPerformingTransition, this.isBreak}) : super(key: key);
+	TaskAppHeader({Key key, @required this.height, @required this.state, this.title, this.text, this.appHeaderWidget, this.helpPage,this.breakPerformingTransition}) : super(key: key);
 
 	@override
 	Size get preferredSize => Size.fromHeight(height);
@@ -37,8 +42,14 @@ class TaskAppHeaderState extends State<TaskAppHeader> with TickerProviderStateMi
 	Animation<Offset> _offsetAnimation;
 	AnimationController _slideController;
 	final String _pageKey = 'page.childSection.taskInProgress';
+	bool _shouldUpdate = false;
+	TimerCubit _timerCompletionCubit;
+	List<int> vibrationPattern = [0, 500, 100, 250, 50, 1000];
+	Timer _updateTimer;
+	final Logger _logger = Logger('TaskAppHeader');
 
-  @override
+
+	@override
   Widget build(BuildContext context) {
 		return Container(
 			color: Colors.grey[50],
@@ -53,9 +64,12 @@ class TaskAppHeaderState extends State<TaskAppHeader> with TickerProviderStateMi
 						helpPage: this.widget.helpPage,
 						isConstrained: true,
 					),
-					_getTimerButtonSection(),
-					_getCurrencyBar(),
-					_getConfettiCanons()
+					if(this.widget.state.taskInstance != null)
+					...[
+						_getTimerButtonSection(),
+						_getCurrencyBar(),
+						_getConfettiCanons()
+					]
 				],
 			)
 		);
@@ -68,17 +82,7 @@ class TaskAppHeaderState extends State<TaskAppHeader> with TickerProviderStateMi
 				child: Row(
 					mainAxisAlignment: MainAxisAlignment.spaceBetween,
 					children: [
-						Align(
-							alignment: Alignment.center,
-							child: BlocProvider<TimerCubit>(
-								create: (_) => TimerCubit(() => 620, CountDirection.down)..startTimer(),
-								child: LargeTimer(
-									textColor: AppColors.darkTextColor,
-									title: '$_pageKey.content.timeLeft',
-									align: alignment,
-								),
-							),
-						),
+						_getTimerSection(),
 						SlideTransition(
 							position: _offsetAnimation,
 							child: _getButtonWidget()
@@ -98,16 +102,22 @@ class TaskAppHeaderState extends State<TaskAppHeader> with TickerProviderStateMi
 					decoration: AppBoxProperties.elevatedContainer.copyWith(borderRadius: BorderRadius.vertical(top: Radius.circular(4.0))),
 					child: Padding(
 						padding: EdgeInsets.symmetric(vertical: 8.0, horizontal: 12.0),
-						child: Row(
+						child: this.widget.state.taskInstance.points != null ? Row(
 							mainAxisAlignment: MainAxisAlignment.spaceBetween,
 							children: [
 								Text(AppLocales.of(context).translate('$_pageKey.content.pointsToGet')),
 								AttributeChip.withCurrency(
-									content: "+30",
-									currencyType: CurrencyType.diamond
+									content: "+" + this.widget.state.taskInstance.points.quantity.toString(),
+									currencyType: this.widget.state.taskInstance.points.type
 								)
 							]
-						)
+						) :
+						Row(
+							mainAxisAlignment: MainAxisAlignment.center,
+						  children: [
+						    Text(AppLocales.of(context).translate('$_pageKey.content.motivate')),
+						  ],
+						),
 					)
 				)
 			)
@@ -142,6 +152,11 @@ class TaskAppHeaderState extends State<TaskAppHeader> with TickerProviderStateMi
 	}
 
 	Widget _getButtonWidget() {
+		if(isBreakNow == null) setState(() {
+			isBreakNow = this.widget.state is TaskInstanceStateBreak;
+			_buttonController = AnimationController(duration: Duration(milliseconds: 450), vsync: this);
+			if(isBreakNow) _buttonController.forward();
+		});
 		return RaisedButton(
 			color: !isBreakNow ? AppColors.childBreakColor : AppColors.childTaskColor,
 			padding: EdgeInsets.zero,
@@ -172,15 +187,73 @@ class TaskAppHeaderState extends State<TaskAppHeader> with TickerProviderStateMi
 					)
 				)
 			),
-			onPressed: () => this.widget.breakPerformingTransition(),
+			onPressed: () => {
+				this.widget.breakPerformingTransition(this.widget.state),
+				this.widget.state is TaskInstanceStateProgress ? _timerCompletionCubit.pauseTimer() : _timerCompletionCubit.resumeTimer()
+			},
 			elevation: 4.0
 		);
 	}
 
+	Widget _getTimerSection() {
+		return Align(
+			alignment: Alignment.center,
+			child: BlocProvider<TimerCubit>(
+				create: _getTimerFun(),
+				child: LargeTimer(
+					textColor: AppColors.darkTextColor,
+					title: _shouldUpdate ? _getTimerTitle() : _getTimerTitle(),
+					align: alignment,
+				),
+			)
+		);
+	}
+
+	void _timeUpdate(Duration duration) {
+		_updateTimer = Timer(duration, () {
+			setState(() {
+			  _shouldUpdate = true;
+			});
+		});
+	}
+
+	CreateBloc _getTimerFun() {
+  	if(_timerCompletionCubit == null) {
+			if(this.widget.state.taskInstance.timer != null) {
+				if(_getTimerInSeconds() - _getDuration() > 0) {
+					int time = _getTimerInSeconds() - _getDuration();
+					_timeUpdate(Duration(seconds: time));
+					_timerCompletionCubit = TimerCubit.down(() => time, true, _onTimerFinish);
+				}
+				else _timerCompletionCubit = TimerCubit.up(() => _getDuration() - _getTimerInSeconds());
+			}
+			else _timerCompletionCubit = TimerCubit.up(() => _getDuration());
+		}
+  	if(this.widget.state is TaskInstanceStateProgress)
+			return (_) => _timerCompletionCubit..startTimer();
+  	else return (_) => _timerCompletionCubit..startTimer(paused: true);
+	}
+
+	int _getDuration() {
+  	return sumDurations(this.widget.state.taskInstance.duration).inSeconds;
+	}
+
+	int _getTimerInSeconds() {
+  	return this.widget.state.taskInstance.timer*60;
+	}
+
+
+	String _getTimerTitle() {
+		if(this.widget.state.taskInstance.timer != null) {
+			if(this.widget.state.taskInstance.timer*60 - sumDurations(this.widget.state.taskInstance.duration).inSeconds > 0)
+				return '$_pageKey.content.timeLeft';
+			else return '$_pageKey.content.latency';
+		}
+		else return '$_pageKey.content.completionTime';
+	}
+
   @override
   void initState() {
-  	isBreakNow = this.widget.isBreak;
-		_buttonController = AnimationController(duration: Duration(milliseconds: 450), vsync: this);
 		_confetti = ConfettiController(
 			duration: Duration(seconds: 10),
 		);
@@ -201,22 +274,39 @@ class TaskAppHeaderState extends State<TaskAppHeader> with TickerProviderStateMi
   	if(!isBreakNow)
 			_buttonController.forward();
   	else 	_buttonController.reverse();
-		isBreakNow = !isBreakNow;
+		setState(() {
+			isBreakNow = !isBreakNow;
+		});
 	}
 
 	void animateSuccess() {
 		_confetti.play();
 	}
 
-	void closeButton() {
+	void onFinish() {
+  	_timerCompletionCubit.pauseTimer();
 		_slideController.forward();
 	}
 
 	@override
   void dispose() {
-  	_confetti.dispose();
+		_updateTimer.cancel();
+		_confetti.dispose();
   	_buttonController.dispose();
   	_slideController.dispose();
     super.dispose();
   }
+
+  void _onTimerFinish() async {
+		try {
+		  if (await Vibration.hasVibrator()) {
+		  	if (await Vibration.hasAmplitudeControl()) {
+		  		Vibration.vibrate(amplitude: 1024, pattern: vibrationPattern);
+		  	}
+		  	else Vibration.vibrate(pattern: vibrationPattern);
+		  }
+		} on Exception catch (e, stacktrace) {
+			_logger.warning("Vibration failed.", e, stacktrace);
+		}
+	}
 }
