@@ -2,18 +2,19 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
+import 'package:fokus_auth/fokus_auth.dart';
 import 'package:logging/logging.dart';
 import 'package:get_it/get_it.dart';
 
-import 'package:fokus/model/auth_user.dart';
 import 'package:fokus/model/db/user/child.dart';
 import 'package:fokus/model/ui/user/ui_user.dart';
 import 'package:fokus/model/db/user/caregiver.dart';
 import 'package:fokus/model/db/user/user_role.dart';
 import 'package:fokus/model/db/user/user.dart';
 import 'package:fokus/services/data/data_repository.dart';
+import 'package:fokus/services/observers/active_user_observer.dart';
 import 'package:fokus/services/plan_keeper_service.dart';
-import 'package:fokus/services/auth/authentication_repository.dart';
+import 'package:fokus/services/notifications/notification_service.dart';
 import 'package:fokus/services/app_config/app_config_repository.dart';
 
 part 'authentication_event.dart';
@@ -22,15 +23,17 @@ part 'authentication_state.dart';
 class AuthenticationBloc extends Bloc<AuthenticationEvent, AuthenticationState> {
 	final Logger _logger = Logger('AuthenticationBloc');
 
-	AuthenticationRepository _authenticationRepository = GetIt.I<AuthenticationRepository>();
-	AppConfigRepository _appConfigRepository = GetIt.I<AppConfigRepository>();
-	DataRepository _dataRepository = GetIt.I<DataRepository>();
-	final PlanKeeperService _outdatedDataService = GetIt.I<PlanKeeperService>();
+	final AuthenticationProvider _authenticationProvider = GetIt.I<AuthenticationProvider>();
+	final AppConfigRepository _appConfigRepository = GetIt.I<AppConfigRepository>();
+	final DataRepository _dataRepository = GetIt.I<DataRepository>();
 
 	StreamSubscription<AuthenticatedUser> _userSubscription;
+	List<ActiveUserObserver> _userObservers = [];
 
   AuthenticationBloc() : super(AuthenticationState.unknown()) {
-	  _userSubscription = _authenticationRepository.user.listen((user) => add(AuthenticationUserChanged(user)));
+	  observeUserChanges(GetIt.I<PlanKeeperService>());
+	  observeUserChanges(GetIt.I<NotificationService>());
+	  _userSubscription = _authenticationProvider.user.listen((user) => add(AuthenticationUserChanged(user)));
   }
 
   @override
@@ -39,17 +42,17 @@ class AuthenticationBloc extends Bloc<AuthenticationEvent, AuthenticationState> 
 		  yield await _processUserChangedEvent(event);
 	  else if (event is AuthenticationChildSignInRequested) {
 			_appConfigRepository.signInChild(event.child.id);
-			_outdatedDataService.onUserSignOut();
 			yield await _signInUser(event.child);
 	  } else if (event is AuthenticationSignOutRequested) {
-	  	_outdatedDataService.onUserSignOut();
+	  	_onUserSignOut(state.user.toDBModel());
 		  if (state.user.role == UserRole.caregiver)
-		    _authenticationRepository.signOut();
+		    _authenticationProvider.signOut();
 		  else {
 			  _appConfigRepository.signOutChild();
 			  add(AuthenticationUserChanged(AuthenticatedUser.empty));
 		  }
-	  }
+	  } else if (event is AuthenticationActiveUserUpdated)
+	  	yield AuthenticationState.authenticated(event.user);
   }
 
 	Future<AuthenticationState> _processUserChangedEvent(AuthenticationUserChanged event) async {
@@ -75,7 +78,7 @@ class AuthenticationBloc extends Bloc<AuthenticationEvent, AuthenticationState> 
   }
 
 	Future<AuthenticationState> _signInUser(User user) async {
-	  await _outdatedDataService.onUserSignIn(user.id, user.role);
+	  _onUserSignIn(user);
 	  return AuthenticationState.authenticated(UIUser.typedFromDBModel(user));
   }
 
@@ -84,4 +87,8 @@ class AuthenticationBloc extends Bloc<AuthenticationEvent, AuthenticationState> 
 		_userSubscription?.cancel();
 		return super.close();
 	}
+
+	void observeUserChanges(ActiveUserObserver observer) => _userObservers.add(observer);
+  void _onUserSignIn(User user) => _userObservers.forEach((observer) => observer.onUserSignIn(user));
+  void _onUserSignOut(User user) => _userObservers.forEach((observer) => observer.onUserSignOut(user));
 }
