@@ -1,33 +1,27 @@
 import 'package:fokus/model/db/date/time_date.dart';
 
 class ChildBadge {
-  String comment;
+  String description;
   TimeDate date;
   int icon;
-  int level;
-  int maxLevel;
   String name;
 
-  ChildBadge({this.comment, this.date, this.icon, this.level, this.maxLevel, this.name});
+  ChildBadge({this.description, this.date, this.icon, this.name});
 
   factory ChildBadge.fromJson(Map<String, dynamic> json) {
     return json != null ? ChildBadge(
-      comment: json['comment'],
+      description: json['description'],
       date: TimeDate.parseDBDate(json['date']),
       icon: json['icon'],
-      level: json['level'],
-      maxLevel: json['maxLevel'],
       name: json['name'],
     ) : null;
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['comment'] = this.comment;
+    data['description'] = this.description;
     data['date'] = this.date.toDBDate();
     data['icon'] = this.icon;
-    data['level'] = this.level;
-    data['maxLevel'] = this.maxLevel;
     data['name'] = this.name;
     return data;
   }
