@@ -1,42 +1,43 @@
+import 'package:equatable/equatable.dart';
+import 'package:fokus/model/db/date/time_date.dart';
 import 'package:fokus/model/db/gamification/badge.dart';
+import 'package:fokus/model/db/gamification/child_badge.dart';
 
-enum UIBadgeMaxLevel { one, three, five }
+class UIBadge extends Equatable {
+	final String name;
+	final String description;
+	final int icon;
 
-extension UIBadgeMaxLevelExtension on UIBadgeMaxLevel {
-  int get value {
-    switch (this) {
-      case UIBadgeMaxLevel.one:
-        return 1;
-      case UIBadgeMaxLevel.three:
-        return 3;
-      case UIBadgeMaxLevel.five:
-        return 5;
-      default:
-        return null;
-    }
-  }
+	UIBadge({this.name, this.description, this.icon = 0});
+	UIBadge.fromDBModel(Badge badge) : this(name: badge.name, description: badge.description, icon: badge.icon);
+
+	UIBadge copyWith({String name, String description, int icon}) {
+		return UIBadge(
+			name: name ?? this.name,
+			description: description ?? this.description,
+			icon: icon ?? this.icon
+		);
+	}
+
+  @override
+  List<Object> get props => [name, description, icon];
 }
 
-class UIBadge {
-	String name;
-	String description;
-	UIBadgeMaxLevel maxLevel;
-	int icon;
-	DateTime date;
+class UIChildBadge extends UIBadge {
+	final TimeDate date;
 
-	UIBadge({
-		this.name,
-		this.description,
-		this.maxLevel = UIBadgeMaxLevel.one,
-		this.icon = 0,
-		this.date
-	});
-	UIBadge.fromDBModel(Badge badge) : this(
-		name: badge.name,
-		description:
-		badge.description,
-		icon: badge.icon,
-		maxLevel: badge.maxLevel == 5 ? UIBadgeMaxLevel.five : (badge.maxLevel == 3 ? UIBadgeMaxLevel.three : UIBadgeMaxLevel.one)
-	);
+	UIChildBadge({String name, String description, int icon, this.date}) : super(name: name, description: description, icon: icon);
+	UIChildBadge.fromDBModel(ChildBadge badge) : this(name: badge.name, description: badge.description, icon: badge.icon, date: badge.date);
+	
+	UIChildBadge copyWith({String name, String description, int icon, DateTime date}) {
+		return UIChildBadge(
+			name: name ?? this.name,
+			description: description ?? this.description,
+			icon: icon ?? this.icon,
+			date: date ?? this.date
+		);
+	}
 
+  @override
+  List<Object> get props => super.props..addAll([date]);
 }
