@@ -1,6 +1,7 @@
 import 'package:bson/bson.dart';
 import 'package:fokus/model/db/date/time_date.dart';
 import 'package:fokus/model/db/date_span.dart';
+import 'package:fokus/model/db/plan/task.dart';
 import 'package:fokus/model/db/plan/task_instance.dart';
 import 'package:fokus/model/db/plan/task_status.dart';
 import 'package:fokus/model/ui/gamification/ui_points.dart';
@@ -12,6 +13,7 @@ enum TaskUIType {completed, available, inBreak, currentlyPerformed, rejected, qu
 extension TaskUITypeGroups on TaskUIType {
 	bool get inProgress => this == TaskUIType.inBreak || this == TaskUIType.currentlyPerformed;
 	bool get wasInProgress => this == TaskUIType.rejected || this == TaskUIType.completed;
+	bool get canBeStarted => this == TaskUIType.available || this == TaskUIType.rejected;
 }
 
 class UITaskInstance extends UITaskBase {
@@ -26,9 +28,10 @@ class UITaskInstance extends UITaskBase {
 	final int Function() elapsedTimePassed;
 
   UITaskInstance({ObjectId id, String name, bool optional, String description, this.timer, this.duration, this.breaks, this.planInstanceId, this.points, this.taskId, this.taskUiType, this.status, this.elapsedTimePassed}) : super(id, name, optional, description);
-	UITaskInstance.singleFromDBModel({TaskInstance task, String name, String description, UIPoints points, int Function() elapsedTimePassed}) : this(id: task.id, name: name, optional: task.optional, description: description, timer: task.timer, duration: task.duration, breaks: task.breaks, planInstanceId: task.planInstanceID, points: points, taskId: task.taskID, taskUiType: TaskInstanceService.getSingleTaskInstanceStatus(task: task), status: task.status, elapsedTimePassed: elapsedTimePassed = _defElapsedTime);
-	UITaskInstance.listFromDBModel({TaskInstance task, String name, String description, UIPoints points, TaskUIType type, int Function() elapsedTimePassed}) : this(id: task.id, name: name, optional: task.optional, description: description, timer: task.timer, duration: task.duration, breaks: task.breaks, planInstanceId: task.planInstanceID, points: points, taskId: task.taskID, taskUiType: type, status: task.status, elapsedTimePassed: elapsedTimePassed = _defElapsedTime);
+	UITaskInstance.singleFromDBModel({TaskInstance taskInstance, String name, String description, UIPoints points, int Function() elapsedTimePassed}) : this(id: taskInstance.id, name: name, optional: taskInstance.optional, description: description, timer: taskInstance.timer, duration: taskInstance.duration, breaks: taskInstance.breaks, planInstanceId: taskInstance.planInstanceID, points: points, taskId: taskInstance.taskID, taskUiType: TaskInstanceService.getSingleTaskInstanceStatus(task: taskInstance), status: taskInstance.status, elapsedTimePassed: elapsedTimePassed = _defElapsedTime);
+	UITaskInstance.singleWithTask({TaskInstance taskInstance, Task task, int Function() elapsedTimePassed}) : this(id: taskInstance.id, name: task.name, optional: taskInstance.optional, description: task.description, timer: taskInstance.timer, duration: taskInstance.duration, breaks: taskInstance.breaks, planInstanceId: taskInstance.planInstanceID, points: task.points != null ? UIPoints(quantity: task.points.quantity, title: task.points.name, createdBy: task.points.createdBy, type: task.points.icon) : null, taskId: taskInstance.taskID, taskUiType: TaskInstanceService.getSingleTaskInstanceStatus(task: taskInstance), status: taskInstance.status, elapsedTimePassed: elapsedTimePassed = _defElapsedTime);
 
+	UITaskInstance.listFromDBModel({TaskInstance taskInstance, String name, String description, UIPoints points, TaskUIType type, int Function() elapsedTimePassed}) : this(id: taskInstance.id, name: name, optional: taskInstance.optional, description: description, timer: taskInstance.timer, duration: taskInstance.duration, breaks: taskInstance.breaks, planInstanceId: taskInstance.planInstanceID, points: points, taskId: taskInstance.taskID, taskUiType: type, status: taskInstance.status, elapsedTimePassed: elapsedTimePassed = _defElapsedTime);
 
 	@override
 	List<Object> get props => super.props..addAll([timer, duration, breaks, planInstanceId, taskId, points]);
