@@ -20,8 +20,10 @@ class UIUser extends Equatable {
 
 	factory UIUser.typedFromDBModel(User user) => user.role == UserRole.caregiver ? UICaregiver.fromDBModel(user) : UIChild.fromDBModel(user);
 
+	User toDBModel() => User(id: id, name: name, role: role, avatar: avatar);
+
 	@override
-  List<Object> get props => [id, name, avatar];
+  List<Object> get props => [id, name, avatar, role];
 
 	@override
   String toString() {
