@@ -1,30 +1,40 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:fokus/logic/task_instance/task_instance_cubit.dart';
 import 'package:fokus/logic/timer/timer_cubit.dart';
+import 'package:fokus/model/ui/plan/ui_plan_instance.dart';
 import 'package:fokus/services/app_locales.dart';
+import 'package:fokus/utils/duration_utils.dart';
 import 'package:fokus/utils/theme_config.dart';
 import 'package:fokus/widgets/buttons/double_circular_notched_button.dart';
 import 'package:fokus/widgets/cards/item_card.dart';
 import 'package:fokus/widgets/cards/sliding_card.dart';
 import 'package:fokus/widgets/chips/attribute_chip.dart';
+import 'package:fokus/widgets/general/app_loader.dart';
 import 'package:fokus/widgets/large_timer.dart';
 import 'package:fokus/widgets/task_app_header.dart';
 import 'package:lottie/lottie.dart';
 
-enum TaskInProgressPageState {currentlyCompleting, inBreak, done, rejected}
 
 class ChildTaskInProgressPage extends StatefulWidget {
+	final UIPlanInstance initialPlanInstance;
+
+  const ChildTaskInProgressPage({Key key, @required this.initialPlanInstance}) : super(key: key);
+
   @override
   _ChildTaskInProgressPageState createState() => _ChildTaskInProgressPageState();
 }
 
 class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with TickerProviderStateMixin {
   final String _pageKey = 'page.childSection.taskInProgress';
-  TaskInProgressPageState _state = TaskInProgressPageState.currentlyCompleting;
 	Animation<Offset> _offsetAnimation;
 	Animation<Offset> _taskListFabAnimation;
 	AnimationController _bottomBarController;
+	bool _isButtonDisabled = false;
+	TimerCubit _timerBreakCubit;
 
 	final GlobalKey<SlidingCardState> _breakCard = GlobalKey<SlidingCardState>();
 	final GlobalKey<SlidingCardState> _completingCard = GlobalKey<SlidingCardState>();
@@ -35,54 +45,66 @@ class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with 
 
 	@override
   Widget build(BuildContext context) {
-  	return Scaffold(
-			appBar: _getHeader(),
-			body: Padding(
-				padding: EdgeInsets.only(bottom: 0.0),
-				child: Stack(
-					children: _getAllCards()
-				)
-			),
-			extendBody: true,
-			bottomNavigationBar: _getBottomNavBar(),
-			floatingActionButtonLocation: FloatingActionButtonLocation.endDocked,
-			floatingActionButton: Row(
-				mainAxisAlignment: MainAxisAlignment.spaceBetween,
-				children: <Widget>[
-					_getFab(true),
-					SlideTransition(
-						position: _taskListFabAnimation,
-					  child: FloatingActionButton.extended(
-					  	label:
-					  		Row(
-					  		  children: [
-					  				Icon(Icons.format_list_bulleted,),
-					  		    Padding(
-					  		      padding: const EdgeInsets.only(left: 8.0),
-					  		      child: Text(AppLocales.of(context).translate('actions.return')),
-					  		    ),
-					  		  ],
-					  		),
-					  	onPressed: () => Navigator.of(context).pop(),
-					  	heroTag: "fabTasks",
-					  	backgroundColor: AppColors.childTaskColor,
-					  ),
+		return BlocBuilder<TaskInstanceCubit, TaskInstanceState>(
+			builder: (context, state) {
+				bool isInitial = false;
+				if(state is TaskInstanceStateInitial) {
+					BlocProvider.of<TaskInstanceCubit>(context).loadTaskInstance();
+					isInitial = true;
+				}
+				return  Scaffold(
+					appBar: isInitial ? _getHeader(TaskInstanceProvider(null, this.widget.initialPlanInstance)) : _getHeader(state),
+					body: isInitial ? Center(child: AppLoader())
+						: Padding(
+						padding: EdgeInsets.only(bottom: 0.0),
+						child: Stack(
+							children: _getAllCards(state)
+						)
 					),
-					_getFab(false)
-				],
-			),
+					extendBody: true,
+					bottomNavigationBar: isInitial ? SizedBox.shrink() : _getBottomNavBar(),
+					floatingActionButtonLocation: FloatingActionButtonLocation.endDocked,
+					floatingActionButton: isInitial ? SizedBox.shrink()
+						: Row(
+						mainAxisAlignment: MainAxisAlignment.spaceBetween,
+						children: <Widget>[
+							_getFab(true, state),
+							SlideTransition(
+								position: _taskListFabAnimation,
+								child: FloatingActionButton.extended(
+									label:
+									Row(
+										children: [
+											Icon(Icons.format_list_bulleted,),
+											Padding(
+												padding: const EdgeInsets.only(left: 8.0),
+												child: Text(AppLocales.of(context).translate('actions.return')),
+											),
+										],
+									),
+									onPressed: () => Navigator.of(context).pop(),
+									heroTag: "fabTasks",
+									backgroundColor: AppColors.childTaskColor,
+								),
+							),
+							_getFab(false, state)
+						],
+					),
+				);
+
+			}
 		);
   }
 
-  Widget _getHeader() {
+  Widget _getHeader(TaskInstanceProvider state) {
 		return TaskAppHeader(
 			height: 330,
 			title: '$_pageKey.header.title',
 			appHeaderWidget: ItemCard(
-				title: "Sprzątanie pokoju",
-				subtitle: "Co każdy poniedziałek, środę, czwartek i piątek",
+				title: state.planInstance.name,
+				subtitle: state.planInstance.description(context),
 				isActive: true,
-				progressPercentage: 0.4,
+				progressPercentage: state.planInstance.completedTaskCount.ceilToDouble() / state.planInstance.taskCount.ceilToDouble(),
 				activeProgressBarColor: AppColors.childActionColor,
 				textMaxLines: 2,
 				chips:
@@ -90,14 +112,14 @@ class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with 
 					AttributeChip.withIcon(
 						icon: Icons.description,
 						color: AppColors.childBackgroundColor,
-						content: AppLocales.of(context).translate('page.childSection.panel.content.taskProgress', {'NUM_TASKS': 2, 'NUM_ALL_TASKS': 5})
+						content: AppLocales.of(context).translate('page.childSection.panel.content.taskProgress', {'NUM_TASKS': state.planInstance.completedTaskCount, 'NUM_ALL_TASKS': state.planInstance.taskCount})
 					)
 				],
 			),
 			helpPage: 'plan_info',
 			key: _header,
 			breakPerformingTransition: _breakPerformingTransition,
-			isBreak: false
+			state: state
 		);
 	}
 
@@ -123,7 +145,7 @@ class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with 
 		);
 	}
 
-  Widget _getFab(bool cancel) {
+  Widget _getFab(bool cancel, state) {
 		return SlideTransition(
 			position: _offsetAnimation,
 			child: Padding(
@@ -132,7 +154,7 @@ class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with 
 					child: Icon(
 						cancel ? Icons.close : Icons.check,
 					),
-					onPressed: cancel ? _onRejection : _onCompletion,
+					onPressed: cancel ? () => _onRejection(state) : () => _onCompletion(state),
 					heroTag: cancel ? "fabClose" :  "fabCheck",
 					backgroundColor: cancel ? AppColors.negativeColor : AppColors.positiveColor,
 				),
@@ -140,16 +162,17 @@ class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with 
 		);
 	}
 
-  List<Widget> _getAllCards() {
+  List<Widget> _getAllCards(TaskInstanceProvider state) {
 		return [
 			SlidingCard(
 				key: _completingCard,
 				cardColor: AppColors.childTaskColor,
 				content: [
 					_getAnimation('assets/animation/jumping_little_man.json'),
-					_getTitle("Tytuł zadania", translate: false),
-					_getSubtitle("Możliwy opcjonalny opis zadania z listą kroków dla dziecka",alignment: TextAlign.justify, translate: false, topPadding: 8),
-				]
+					_getTitle(state.taskInstance.name, translate: false),
+					if(state.taskInstance.description != null) _getSubtitle(state.taskInstance.description ,alignment: TextAlign.justify, translate: false, topPadding: 8),
+				],
+				showFirst: state is TaskInstanceStateProgress,
 			),
 			SlidingCard(
 				key: _breakCard,
@@ -160,15 +183,16 @@ class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with 
 					Padding(
 						padding: const EdgeInsets.symmetric(horizontal: 8.0),
 						child: BlocProvider<TimerCubit>(
-							create: (_) => TimerCubit(() => 0)..startTimer(),
+							create: _getTimerBreakCubit(state),
 							child:LargeTimer(
 								textColor: AppColors.lightTextColor,
 								title: '$_pageKey.content.breakTime',
 								align: CrossAxisAlignment.center,
 							),
-						),
+						)
 					)
-				]
+				],
+				showFirst: state is TaskInstanceStateBreak,
 			),
 			SlidingCard(
 				key: _rejectCard,
@@ -207,6 +231,14 @@ class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with 
 			height: 100,
 			child: Lottie.asset(data)
 		);
+	}
+
+	CreateBloc _getTimerBreakCubit(state) {
+		if(_timerBreakCubit == null) {
+			_timerBreakCubit = TimerCubit.up(() => sumDurations(state.taskInstance.breaks).inSeconds);
+		}
+		if(state is TaskInstanceStateProgress) return (_) => _timerBreakCubit..startTimer(paused: true);
+		else return (_) => _timerBreakCubit..startTimer();
 	}
 
 	Widget _getTitle(String title, {bool translate = true}) {
@@ -249,57 +281,54 @@ class _ChildTaskInProgressPageState extends State<ChildTaskInProgressPage> with 
 			parent: _bottomBarController,
 			curve: Curves.easeInOutQuint,
 		));
-		WidgetsBinding.instance.addPostFrameCallback((_) {
-			_completingCard.currentState.jumpToEnd();
-			_breakCard.currentState.jumpToEnd();
-			_doneCard.currentState.jumpToEnd();
-			_rejectCard.currentState.jumpToEnd();
-			if(_state == TaskInProgressPageState.currentlyCompleting)
-				_completingCard.currentState.openCard();
-			else _breakCard.currentState.openCard();
-		});
 	}
 
- void _breakPerformingTransition() {
-  	setState(() {
-  		if(_state == TaskInProgressPageState.currentlyCompleting) {
-  			_completingCard.currentState.closeCard();
+ void _breakPerformingTransition(state) {
+  	if(!_isButtonDisabled) setState(() {
+  		_isButtonDisabled = true;
+  		if(state is TaskInstanceStateProgress) {
+				BlocProvider.of<TaskInstanceCubit>(context).switchToBreak();
+				_completingCard.currentState.closeCard();
   			_breakCard.currentState.openCard();
-  			_state = TaskInProgressPageState.inBreak;
   			this._header.currentState.animateButton();
+  			_timerBreakCubit.resumeTimer();
 			}
-  		else if(_state == TaskInProgressPageState.inBreak) {
+  		else if(state is TaskInstanceStateBreak) {
+				BlocProvider.of<TaskInstanceCubit>(context).switchToProgress();
 				_breakCard.currentState.closeCard();
 				_completingCard.currentState.openCard();
-				_state = TaskInProgressPageState.currentlyCompleting;
 				this._header.currentState.animateButton();
+				_timerBreakCubit.pauseTimer();
 			}
+			Timer(Duration(seconds: 2), () {
+				_isButtonDisabled = false;
+			});
   	});
   }
 
-  void _onCompletion() {
-		if(_state != TaskInProgressPageState.rejected && _state != TaskInProgressPageState.done)
+  void _onCompletion(state) {
+		if(state is! TaskInstanceStateRejected && state is! TaskInstanceStateDone)
 			setState(() {
-				_closeWidgetsOnFinish();
-				_state = TaskInProgressPageState.done;
+				BlocProvider.of<TaskInstanceCubit>(context).markAsDone();
+				_closeWidgetsOnFinish(state);
 				_doneCard.currentState.openCard();
 				this._header.currentState.animateSuccess();
-				this._header.currentState.closeButton();
+				this._header.currentState.onFinish();
 			});
 	}
 
-	void _onRejection() {
-		if(_state != TaskInProgressPageState.rejected && _state != TaskInProgressPageState.done)
+	void _onRejection(state) {
+		if(state is! TaskInstanceStateRejected && state is! TaskInstanceStateDone)
 			setState(() {
-				_closeWidgetsOnFinish();
-				_state = TaskInProgressPageState.rejected;
+				BlocProvider.of<TaskInstanceCubit>(context).markAsRejected();
+				_closeWidgetsOnFinish(state);
 				_rejectCard.currentState.openCard();
-				this._header.currentState.closeButton();
+				this._header.currentState.onFinish();
 			});
 	}
 
-	void _closeWidgetsOnFinish() {
-		if(_state == TaskInProgressPageState.currentlyCompleting)
+	void _closeWidgetsOnFinish(state) {
+		if(state is TaskInstanceStateProgress)
 			_completingCard.currentState.closeCard();
 		else _breakCard.currentState.closeCard();
 		_bottomBarController.forward();
