@@ -1,4 +1,5 @@
 import 'package:fokus/services/data/award/award_db_repository.dart';
+import 'package:fokus/services/data/currency/currency_db_repository.dart';
 
 import 'db_repository.dart';
 import 'mongodb_provider.dart';
@@ -9,7 +10,7 @@ import '../user/user_db_repository.dart';
 import '../data_repository.dart';
 import '../task/task_db_repository.dart';
 
-class DbDataRepository with UserDbRepository, PlanDbRepository, TaskDbRepository, AwardDbRepository implements DataRepository, DbRepository {
+class DbDataRepository with UserDbRepository, PlanDbRepository, TaskDbRepository, AwardDbRepository, CurrencyDbRepository implements DataRepository, DbRepository {
 	@override
 	final MongoDbProvider dbClient = MongoDbProvider();
 
