@@ -4,7 +4,7 @@ import 'package:fokus/model/db/user/user_role.dart';
 import 'package:mongo_dart/mongo_dart.dart';
 
 abstract class UserRepository {
-	Future<User> getUser({ObjectId id, ObjectId connected, String authenticationId, UserRole role, List<String> fields});
+	Future<User> getUser({ObjectId id, ObjectId connected, String authenticationId, String notificationId, UserRole role, List<String> fields});
 	Future<List<User>> getUsers({List<ObjectId> ids, ObjectId connected, UserRole role, List<String> fields});
 
 	Future<Map<ObjectId, String>> getUserNames(List<ObjectId> users);
@@ -12,4 +12,7 @@ abstract class UserRepository {
 
 	Future createUser(User user);
 	Future updateUser(ObjectId userId, {List<ObjectId> newConnections, List<Points> points});
+
+	Future insertNotificationID(ObjectId userId, String notificationId);
+	Future removeNotificationID(String notificationID, {ObjectId userId});
 }
