@@ -290,15 +290,19 @@ class AppHeader extends StatelessWidget {
 		);
 	 }
 
-	 Widget _userHeaderIcons(BuildContext context) {
+	Widget _userHeaderIcons(BuildContext context) {
+		var auth = context.bloc<AuthenticationBloc>();
 		return Row(
 			children: <Widget>[
-				headerIconButton(Icons.notifications, () => Navigator.of(context).pushNamed(AppPage.notificationsPage.name)),
+				// Supporting system notification only for now
+				// headerIconButton(Icons.notifications, () => Navigator.of(context).pushNamed(AppPage.notificationsPage.name)),
 				PopupMenuList(
 					lightTheme: true,
 					items: [
-						UIButton('actions.signOut', () => context.bloc<AuthenticationBloc>().add(AuthenticationSignOutRequested())),
 						UIButton('navigation.settings', () => Navigator.of(context).pushNamed(AppPage.settingsPage.name)),
+						if(auth.state.user.role == UserRole.caregiver)
+							UIButton('navigation.caregiver.currencies', () => Navigator.of(context).pushNamed(AppPage.caregiverCurrencies.name)),
+						UIButton('actions.signOut', () => auth.add(AuthenticationSignOutRequested())),
 					]
 				),
 			],
